requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
qrcode>=7.0.0
pyzbar>=0.1.8
scikit-learn>=0.24.0